feat: migrate loader, appLoader to TS
